Fix shared/sql: update scripts to GN v2.6.1, add utils functions

+++ shared/data/sql/synthese_after_delete.sql
@@ -11,31 +11,41 @@ SET search_path = gn_synthese, public, pg_catalog ;
 
 
 \echo '----------------------------------------------------------------------------'
-\echo 'Replay trigger "tri_del_area_synt_maj_corarea_tax" actions'
+\echo 'For GeoNature < v2.6.0, replay trigger "tri_del_area_synt_maj_corarea_tax" actions'
+DO $$
+    BEGIN
+        IF EXISTS (
+            SELECT 1
+            FROM pg_trigger
+            WHERE tgname = 'tri_del_area_synt_maj_corarea_tax'
+        ) IS TRUE THEN
+            RAISE NOTICE ' Deleting cor_area_synthese entries not linked with synthese' ;
+            DELETE FROM cor_area_synthese AS cas
+            WHERE NOT EXISTS (
+                SELECT 'X'
+                FROM synthese AS s
+                WHERE s.id_synthese = cas.id_synthese
+            ) ;
 
-\echo ' Deleting cor_area_synthese entries not linked with synthese'
-DELETE FROM cor_area_synthese AS cas
-WHERE NOT EXISTS (
-    SELECT 'X'
-    FROM synthese AS s
-    WHERE s.id_synthese = cas.id_synthese
-) ;
+            RAISE NOTICE ' Clean table cor_area_taxon' ;
+            TRUNCATE TABLE cor_area_taxon ;
+            -- TO AVOID TRUNCATE : add condition on id_source or id_dataset to reduce synthese table entries in below insert
 
-\echo ' Clean table cor_area_taxon'
-TRUNCATE TABLE cor_area_taxon ;
--- TO AVOID TRUNCATE : add condition on id_source or id_dataset to reduce synthese table entries in below insert
-
-\echo ' Reinsert all data in cor_area_taxon'
-INSERT INTO cor_area_taxon (id_area, cd_nom, last_date, nb_obs)
-    SELECT cor.id_area, s.cd_nom, MAX(s.date_min) AS last_date, COUNT(s.id_synthese) AS nb_obs
-    FROM cor_area_synthese AS cor
-        JOIN synthese AS s
-            ON (s.id_synthese = cor.id_synthese)
-    GROUP BY cor.id_area, s.cd_nom ;
-
+            RAISE NOTICE ' Reinsert all data in cor_area_taxon' ;
+            INSERT INTO cor_area_taxon (id_area, cd_nom, last_date, nb_obs)
+                SELECT cor.id_area, s.cd_nom, MAX(s.date_min) AS last_date, COUNT(s.id_synthese) AS nb_obs
+                FROM cor_area_synthese AS cor
+                    JOIN synthese AS s
+                        ON (s.id_synthese = cor.id_synthese)
+                GROUP BY cor.id_area, s.cd_nom ;
+        ELSE
+      		RAISE NOTICE ' GeoNature > v2.6.0 => trigger "tri_del_area_synt_maj_corarea_tax" not exists !' ;
+        END IF ;
+    END
+$$ ;
 
 \echo '-------------------------------------------------------------------------------'
-\echo 'For GeoNature < v2.3.2 replay and enable synthese trigger "trg_refresh_taxons_forautocomplete"'
+\echo 'For GeoNature < v2.3.2, replay and enable synthese trigger "trg_refresh_taxons_forautocomplete"'
 DO $$
     BEGIN
         IF EXISTS (
@@ -61,8 +71,20 @@ DO $$
 $$ ;
 
 \echo '----------------------------------------------------------------------------'
-\echo 'Enable trigger "tri_del_area_synt_maj_corarea_tax"'
-ALTER TABLE synthese ENABLE TRIGGER tri_del_area_synt_maj_corarea_tax ;
+\echo 'For GeoNature < v2.6.0, enable trigger "tri_del_area_synt_maj_corarea_tax"'
+DO $$
+    BEGIN
+        IF EXISTS (
+            SELECT 1
+            FROM pg_trigger
+            WHERE tgname = 'tri_del_area_synt_maj_corarea_tax'
+        ) IS TRUE THEN
+            ALTER TABLE synthese ENABLE TRIGGER tri_del_area_synt_maj_corarea_tax ;
+        ELSE
+      		RAISE NOTICE ' GeoNature > v2.6.0 => trigger "tri_del_area_synt_maj_corarea_tax" not exists !' ;
+        END IF ;
+    END
+$$ ;
 
 \echo '-------------------------------------------------------------------------------'
 \echo 'COMMIT if all is ok:'
